requirements.txt
click==8.1.3
cloudpickle==2.2.0
# crafter==1.8.1
# d4rl==1.1
dm_control==1.0.3
gym==0.21.0
matplotlib==3.6.1
# metaworld==0.0.1.dev0
mujoco_py==2.1.2.14
numpy==1.19.5
pandas==1.4.4
pillow==9.2.0
rich==12.0.1
ruamel.base==1.0.0
tensorflow==2.7.0
tensorflow_probability==0.14.1
termcolor==1.1.0
